move components out into own files
