+++ styles/pages/_ClassPage.scss
@@ -87,7 +87,7 @@
     }
   }
 
-  .classPageInfoHeader {
+  .pageContentHeader {
     margin-bottom: 10px;
   }
